+++ semcore/bulk-textarea/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.4.0] - 2025-08-29
+
+### Added
+
+- Controlled mode for errors.
+- `onImmediatelyChange` internal property.
+
 ## [16.3.0] - 2025-08-07
 
 ### Changed